Move fix_array_dtype to openquake/server (impacting only on web api) [demos]
